Add Helper for Reading Files
Add an enum class that stores the type of file to read, along with a fileToStream function that takes the day and enum. This allows for much simpler code on the callers end for loading the data from a file.

To help with when we need to read the whole file, without it being a stream, add a fileToString function that takes either the fileName, or the day and enum