Performance Optimizations
Avoid using what effectively amounts to Pair<Int, Int> to store the rules, and instead store them in a Map<Int, Set<Int>> where the left-hand-side is the key, and the right-hand-side(s) are the values in the set. This allows for faster and simpler order checking and sorting, because we only need to do 2 hash checks in the worst case.