Add Benchmarking Util
Create a generic function to help with benchmarking code. It allows for testing with and without JIT warmup.
Technically, it would be better to have an interface that all Advent of Code classes implement, however that's a LARGE refactor for prior years. There are also edge cases in the input where the part1/part2 functions need to take a value to determine testing/actual input version. This is a low-impact addition to still be able to test