AoC 2024 Day 15 Optimization
Reduce intermediate maps by creating mutable copies as late as possible.
In addition, optimize for the general case of just moving the cursor. This is done by not keeping the cursor as a char in the Warehouse grid, and allows for us to simply make a shallow copy of the object with the new position. This reduces the runtime from 1.2s to 123ms