Add Shared Code for Converting a String -> Map<Point2D, T>
There are a number of problems that devolve into converting a string with flatMapIndexed & mapIndexed to build up a row-colum map. At this point, there should be shared code for generating this. To make it easier to have the value of the map be any value, add overload that takes a transformation function, and one that removes null values.
This was not added while reading input to make it easier to use for non-DataFile strings