Create AoC Interface for Classes to Implement
Create an interface that expects functions to implement a 'part1' and 'part2' function. This allows for easier implementation of the 'report', and 'benchmark' functions to run a given day.
Each of the existing Year/Day classes that exist have been updated to use this format. Any parts that had to take an argument as the value, have instead had that moved into the primary constructor, when possible.